+++ messages/de/pages.php
@@ -70,4 +70,8 @@ return [
     'documents_download_all' => 'Alle herunterladen',
     'documents_none_yet' => 'Es wurden noch keine Dokumente veröffentlicht.',
     'documents_sort' => 'Dokumentengruppen anordnen',
+
+    'err_permission' => 'Keine Berechtigung, diese Seite zu bearbeiten',
+    'err_exists_site' => 'Es gibt bereits eine globale Seite mit dieser URL.',
+    'err_exists_con' => 'Es gibt bereits eine Seite mit dieser URL.',
 ];